// openi40.ui/openi40-scheduling-ui/projects/openi40-scheduler-api/src/lib/model/purchaseSupplyDto.ts
/**
 * 
 * This code is part of the OpenI40 open source advanced production scheduler
 * platform suite, have look to its licencing options.
 * Web site: http://openi40.org/  
 * Github: https://github.com/openi40/OpenI40Platform
 * We hope you enjoy implementing new amazing projects with it.
 * @author architectures@openi40.org
 *
 */
/**
 * Openi40
 * The open source industy 4.0 production scheduler & MES platform
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SupplyReservationDto } from './supplyReservationDto';


export interface PurchaseSupplyDto { 
    availabilityDateTime?: Date;
    availabilityWeek?: number;
    code?: string;
    description?: string;
    id?: string;
    productCode?: string;
    qtyAvailable?: number;
    qtyReserved?: number;
    qtyTotal?: number;
    reservations?: Array<SupplyReservationDto>;
    warehouseCode?: string;
}